requirements.txt
Flask
python-dotenv
google-cloud-translate
google-ai-generativelanguage
google-cloud-firestore
google-auth
